marshal and unmarshall pointer
