--- api/v1beta1/vcdcluster_webhook.go
@@ -1,57 +1,0 @@
-/*
-Copyright 2021.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package v1beta1
-
-import (
-	"k8s.io/apimachinery/pkg/runtime"
-	ctrl "sigs.k8s.io/controller-runtime"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/webhook"
-)
-
-// log is for logging in this package.
-var vcdclusterlog = logf.Log.WithName("vcdcluster-resource")
-
-func (r *VCDCluster) SetupWebhookWithManager(mgr ctrl.Manager) error {
-	return ctrl.NewWebhookManagedBy(mgr).
-		For(r).
-		Complete()
-}
-
-// +kubebuilder:webhook:verbs=create;update;delete,path=/validate-infrastructure-cluster-x-k8s-io-v1beta1-vcdcluster,mutating=false,failurePolicy=fail,matchPolicy=Equivalent,groups=infrastructure.cluster.x-k8s.io,resources=vcdclusters,versions=v1beta1,name=validation.vcdcluster.infrastructure.cluster.x-k8s.io,sideEffects=None,admissionReviewVersions=v1;v1beta1
-// +kubebuilder:webhook:verbs=create;update,path=/mutate-infrastructure-cluster-x-k8s-io-v1beta1-vcdcluster,mutating=true,failurePolicy=fail,matchPolicy=Equivalent,groups=infrastructure.cluster.x-k8s.io,resources=vcdclusters,versions=v1beta1,name=default.vcdcluster.infrastructure.cluster.x-k8s.io,sideEffects=None,admissionReviewVersions=v1;v1beta1
-
-var _ webhook.Validator = &VCDCluster{}
-var _ webhook.Defaulter = &VCDCluster{}
-
-// Default implements webhook.Defaulter so a webhook will be registered for the type.
-func (c *VCDCluster) Default() {
-}
-
-// ValidateCreate implements webhook.Validator so a webhook will be registered for the type.
-func (c *VCDCluster) ValidateCreate() error {
-	return nil
-}
-
-func (c *VCDCluster) ValidateUpdate(oldRaw runtime.Object) error {
-	return nil
-}
-
-// ValidateDelete implements webhook.Validator so a webhook will be registered for the type.
-func (c *VCDCluster) ValidateDelete() error {
-	return nil
-}
