+++ config/crd/kustomization.yaml
@@ -1,6 +1,5 @@
 commonLabels:
-  cluster.x-k8s.io/v1beta2: v1beta2
-  cluster.x-k8s.io/v1beta1: v1beta1
+  cluster.x-k8s.io/v1beta1: v1beta2
   cluster.x-k8s.io/v1alpha4: v1alpha4
   clusterctl.cluster.x-k8s.io: ""
   clusterctl.cluster.x-k8s.io/move: ""
