+++ config/webhook/manifests.yaml
@@ -8,12 +8,12 @@ metadata:
 webhooks:
 - admissionReviewVersions:
   - v1
-  - v1beta1
+  - v1beta2
   clientConfig:
     service:
       name: webhook-service
       namespace: system
-      path: /mutate-infrastructure-cluster-x-k8s-io-v1beta1-vcdcluster
+      path: /mutate-infrastructure-cluster-x-k8s-io-v1beta2-vcdcluster
   failurePolicy: Fail
   matchPolicy: Equivalent
   name: default.vcdcluster.infrastructure.cluster.x-k8s.io
@@ -21,7 +21,7 @@ webhooks:
   - apiGroups:
     - infrastructure.cluster.x-k8s.io
     apiVersions:
-    - v1beta1
+    - v1beta2
     operations:
     - CREATE
     - UPDATE
@@ -58,12 +58,12 @@ metadata:
 webhooks:
 - admissionReviewVersions:
   - v1
-  - v1beta1
+  - v1beta2
   clientConfig:
     service:
       name: webhook-service
       namespace: system
-      path: /validate-infrastructure-cluster-x-k8s-io-v1beta1-vcdcluster
+      path: /validate-infrastructure-cluster-x-k8s-io-v1beta2-vcdcluster
   failurePolicy: Fail
   matchPolicy: Equivalent
   name: validation.vcdcluster.infrastructure.cluster.x-k8s.io
@@ -71,7 +71,7 @@ webhooks:
   - apiGroups:
     - infrastructure.cluster.x-k8s.io
     apiVersions:
-    - v1beta1
+    - v1beta2
     operations:
     - CREATE
     - UPDATE
